requirements.txt
django==3.0.4
djangorestframework==3.10.3
django-rest-swagger==2.2.0
drf-yasg==1.17.0
drf-generators==0.5.0
